+++ src/engine/editor/sequence/SequenceEditorController.cpp
@@ -24,7 +24,7 @@ namespace Unnamed {
 		}
 
 		[[nodiscard]] int64_t QuantizeFrame(const float frame) {
-			return static_cast<int64_t>(std::llround(std::max(0.0f, frame)));
+			return std::llround(std::max(0.0f, frame));
 		}
 
 		void EnsureCurveChannelId(SequenceRichCurveAssetData& ioCurve) {
@@ -60,14 +60,19 @@ namespace Unnamed {
 			);
 			std::ranges::sort(
 				ioCurve.keys,
-				[](const SequenceFloatKeyAssetData& lhs, const SequenceFloatKeyAssetData& rhs) {
+				[](
+				const SequenceFloatKeyAssetData& lhs,
+				const SequenceFloatKeyAssetData& rhs
+			) {
 					return lhs.frame < rhs.frame;
 				}
 			);
 		}
 	}
 
-	void SequenceEditorController::Initialize(World* world, AssetManager* assetManager) {
+	void SequenceEditorController::Initialize(
+		World* world, AssetManager* assetManager
+	) {
 		mWorld        = world;
 		mAssetManager = assetManager;
 		EnsurePreviewPlayer();
@@ -77,7 +82,8 @@ namespace Unnamed {
 			mAssetManager->RegisterReload(
 				[](const AssetID assetId) {
 					if (gSequenceEditorReloadReceiver) {
-						gSequenceEditorReloadReceiver->HandleAssetReload(assetId);
+						gSequenceEditorReloadReceiver->HandleAssetReload(
+							assetId);
 					}
 				}
 			);
@@ -90,7 +96,7 @@ namespace Unnamed {
 			return;
 		}
 
-		mWorld = world;
+		mWorld             = world;
 		mPreviewRegistered = false;
 		if (mWorld && mPreviewPlayer && !mPreviewRegistered) {
 			mWorld->GetSequenceRuntime().RegisterPlayer(mPreviewPlayer);
@@ -136,7 +142,8 @@ namespace Unnamed {
 
 		auto document = std::make_shared<SequenceEditorDocument>(path);
 		if (!document->ReloadFromDisk()) {
-			Warning("SequenceEditor", "Failed to load sequence document: {}", path);
+			Warning("SequenceEditor", "Failed to load sequence document: {}",
+			        path);
 			return false;
 		}
 
@@ -215,7 +222,8 @@ namespace Unnamed {
 		return mDocuments[mActiveDocumentIndex].get();
 	}
 
-	const SequenceEditorDocument* SequenceEditorController::GetActiveDocument() const {
+	const SequenceEditorDocument*
+	SequenceEditorController::GetActiveDocument() const {
 		return const_cast<SequenceEditorController*>(this)->GetActiveDocument();
 	}
 
@@ -237,8 +245,10 @@ namespace Unnamed {
 		mPlayheadFrame       = 0.0f;
 
 		if (const SequenceEditorDocument* document = GetActiveDocument()) {
-			mAutoKeyEnabled  = document->GetAuthoringData().editor.autoKeyEnabled;
-			mScrubFireEvents = document->GetAuthoringData().editor.scrubFireEvents;
+			mAutoKeyEnabled = document->GetAuthoringData().editor.
+			                            autoKeyEnabled;
+			mScrubFireEvents = document->GetAuthoringData().editor.
+			                             scrubFireEvents;
 		}
 		RefreshPreviewAsset();
 	}
@@ -247,7 +257,8 @@ namespace Unnamed {
 		return mSelection;
 	}
 
-	const SequenceEditorSelection& SequenceEditorController::GetSelection() const {
+	const SequenceEditorSelection&
+	SequenceEditorController::GetSelection() const {
 		return mSelection;
 	}
 
@@ -260,14 +271,16 @@ namespace Unnamed {
 		const bool  seekNow
 	) {
 		SequenceEditorDocument* document = GetActiveDocument();
-		const float maxFrame = document ?
-			                       static_cast<float>(
-				                       std::max<int64_t>(
-					                       0,
-					                       document->GetAuthoringData().lengthFrames
-				                       )
-			                       ) :
-			                       0.0f;
+		const float             maxFrame = document ?
+			                                   static_cast<float>(
+				                                   std::max<int64_t>(
+					                                   0,
+					                                   document->
+					                                   GetAuthoringData().
+					                                   lengthFrames
+				                                   )
+			                                   ) :
+			                                   0.0f;
 		mPlayheadFrame = std::clamp(frame, 0.0f, std::max(0.0f, maxFrame));
 
 		if (!seekNow || !mPreviewPlayer || !mWorld) {
@@ -288,7 +301,8 @@ namespace Unnamed {
 		if (!mPreviewPlayer) {
 			return;
 		}
-		mPreviewPlayer->SetPlaybackDirection(SEQUENCE_PLAYBACK_DIRECTION::FORWARD);
+		mPreviewPlayer->SetPlaybackDirection(
+			SEQUENCE_PLAYBACK_DIRECTION::FORWARD);
 		mPreviewPlayer->SetSeekEventPolicy(
 			mScrubFireEvents ?
 				SEQUENCE_SEEK_EVENT_POLICY::FIRE_IN_RANGE :
@@ -302,7 +316,8 @@ namespace Unnamed {
 		if (!mPreviewPlayer) {
 			return;
 		}
-		mPreviewPlayer->SetPlaybackDirection(SEQUENCE_PLAYBACK_DIRECTION::BACKWARD);
+		mPreviewPlayer->SetPlaybackDirection(
+			SEQUENCE_PLAYBACK_DIRECTION::BACKWARD);
 		mPreviewPlayer->SetSeekEventPolicy(
 			mScrubFireEvents ?
 				SEQUENCE_SEEK_EVENT_POLICY::FIRE_IN_RANGE :
@@ -327,7 +342,8 @@ namespace Unnamed {
 	}
 
 	bool SequenceEditorController::IsPreviewPlaying() const {
-		return mPreviewPlayer && mPreviewPlayer->GetState() == SEQUENCE_PLAYER_STATE::PLAYING;
+		return mPreviewPlayer && mPreviewPlayer->GetState() ==
+		       SEQUENCE_PLAYER_STATE::PLAYING;
 	}
 
 	bool SequenceEditorController::IsAutoKeyEnabled() const {
@@ -345,7 +361,9 @@ namespace Unnamed {
 		return mScrubFireEvents;
 	}
 
-	void SequenceEditorController::SetScrubFireEventsEnabled(const bool enabled) {
+	void SequenceEditorController::SetScrubFireEventsEnabled(
+		const bool enabled
+	) {
 		mScrubFireEvents = enabled;
 		if (SequenceEditorDocument* document = GetActiveDocument()) {
 			document->GetAuthoringData().editor.scrubFireEvents = enabled;
@@ -368,17 +386,17 @@ namespace Unnamed {
 	}
 
 	void SequenceEditorController::OnGizmoTransformChanged(
-		const uint64_t          entityGuid,
-		const Vec3&             position,
-		const Quaternion&       rotation,
-		const Vec3&             scale
+		const uint64_t    entityGuid,
+		const Vec3&       position,
+		const Quaternion& rotation,
+		const Vec3&       scale
 	) {
 		if (!mAutoKeyEnabled || entityGuid == 0) {
 			return;
 		}
 
 		const int64_t keyFrame = QuantizeFrame(mPlayheadFrame);
-		const bool modified = ModifyActiveDocument(
+		const bool    modified = ModifyActiveDocument(
 			[&](SequenceAuthoringData& data) {
 				int32_t trackIndex   = -1;
 				int32_t sectionIndex = -1;
@@ -403,22 +421,31 @@ namespace Unnamed {
 					return;
 				}
 
-				SequenceSectionAssetData& section = track.sections[sectionIndex];
-				AddOrUpdateFloatKey(section.transformPosX, keyFrame, position.x);
-				AddOrUpdateFloatKey(section.transformPosY, keyFrame, position.y);
-				AddOrUpdateFloatKey(section.transformPosZ, keyFrame, position.z);
-				AddOrUpdateFloatKey(section.transformRotX, keyFrame, rotation.x);
-				AddOrUpdateFloatKey(section.transformRotY, keyFrame, rotation.y);
-				AddOrUpdateFloatKey(section.transformRotZ, keyFrame, rotation.z);
-				AddOrUpdateFloatKey(section.transformRotW, keyFrame, rotation.w);
+				SequenceSectionAssetData& section = track.sections[
+					sectionIndex];
+				AddOrUpdateFloatKey(section.transformPosX, keyFrame,
+				                    position.x);
+				AddOrUpdateFloatKey(section.transformPosY, keyFrame,
+				                    position.y);
+				AddOrUpdateFloatKey(section.transformPosZ, keyFrame,
+				                    position.z);
+				AddOrUpdateFloatKey(section.transformRotX, keyFrame,
+				                    rotation.x);
+				AddOrUpdateFloatKey(section.transformRotY, keyFrame,
+				                    rotation.y);
+				AddOrUpdateFloatKey(section.transformRotZ, keyFrame,
+				                    rotation.z);
+				AddOrUpdateFloatKey(section.transformRotW, keyFrame,
+				                    rotation.w);
 				AddOrUpdateFloatKey(section.transformScaleX, keyFrame, scale.x);
 				AddOrUpdateFloatKey(section.transformScaleY, keyFrame, scale.y);
 				AddOrUpdateFloatKey(section.transformScaleZ, keyFrame, scale.z);
 
-				data.lengthFrames = std::max(data.lengthFrames, keyFrame);
+				data.lengthFrames       = std::max(data.lengthFrames, keyFrame);
 				mSelection.trackIndex   = trackIndex;
 				mSelection.sectionIndex = sectionIndex;
-				mSelection.floatChannel = SEQUENCE_EDITOR_FLOAT_CHANNEL::TRANSFORM_POS_X;
+				mSelection.floatChannel =
+					SEQUENCE_EDITOR_FLOAT_CHANNEL::TRANSFORM_POS_X;
 			}
 		);
 		if (modified) {
@@ -439,18 +466,22 @@ namespace Unnamed {
 			return;
 		}
 
-		mAutoKeyEnabled  = activeDocument->GetAuthoringData().editor.autoKeyEnabled;
-		mScrubFireEvents = activeDocument->GetAuthoringData().editor.scrubFireEvents;
+		mAutoKeyEnabled = activeDocument->GetAuthoringData().editor.
+		                                  autoKeyEnabled;
+		mScrubFireEvents = activeDocument->GetAuthoringData().editor.
+		                                   scrubFireEvents;
 		RefreshPreviewAsset();
 	}
 
 	void SequenceEditorController::EnsurePreviewPlayer() {
 		if (!mPreviewPlayer) {
 			mPreviewPlayer = std::make_shared<SequencePlayer>();
-			mPreviewPlayer->SetCompletionMode(SEQUENCE_COMPLETION_MODE::KEEP_STATE);
+			mPreviewPlayer->SetCompletionMode(
+				SEQUENCE_COMPLETION_MODE::KEEP_STATE);
 			mPreviewPlayer->SetLoop(false);
 			mPreviewPlayer->SetPlayRate(1.0f);
-			mPreviewPlayer->SetPlaybackDirection(SEQUENCE_PLAYBACK_DIRECTION::FORWARD);
+			mPreviewPlayer->SetPlaybackDirection(
+				SEQUENCE_PLAYBACK_DIRECTION::FORWARD);
 		}
 		if (mWorld && !mPreviewRegistered) {
 			mWorld->GetSequenceRuntime().RegisterPlayer(mPreviewPlayer);
@@ -469,9 +500,10 @@ namespace Unnamed {
 			return;
 		}
 
-		const AssetID previewAssetId = activeDocument->RebuildPreviewRuntimeAsset(
-			*mAssetManager
-		);
+		const AssetID previewAssetId = activeDocument->
+			RebuildPreviewRuntimeAsset(
+				*mAssetManager
+			);
 		if (previewAssetId == kInvalidAssetID) {
 			return;
 		}
@@ -513,11 +545,11 @@ namespace Unnamed {
 		}
 		if (bindingId == 0) {
 			SequenceBindingAssetData binding = {};
-			binding.bindingId = AllocateStableId();
-			binding.entityGuid = entityGuid;
-			binding.componentStableName = "engine.Transform";
-			binding.propertyPath = "";
-			bindingId = binding.bindingId;
+			binding.bindingId                = AllocateStableId();
+			binding.entityGuid               = entityGuid;
+			binding.componentStableName      = "engine.Transform";
+			binding.propertyPath             = "";
+			bindingId                        = binding.bindingId;
 			ioData.bindings.emplace_back(std::move(binding));
 		}
 
@@ -525,7 +557,8 @@ namespace Unnamed {
 			mSelection.trackIndex >= 0 &&
 			mSelection.trackIndex < static_cast<int32_t>(ioData.tracks.size())
 		) {
-			const SequenceTrackAssetData& selectedTrack = ioData.tracks[mSelection.trackIndex];
+			const SequenceTrackAssetData& selectedTrack = ioData.tracks[
+				mSelection.trackIndex];
 			if (
 				selectedTrack.trackType == SEQUENCE_TRACK_TYPE::TRANSFORM &&
 				selectedTrack.bindingId == bindingId
@@ -535,7 +568,8 @@ namespace Unnamed {
 		}
 
 		if (outTrackIndex < 0) {
-			for (int32_t i = 0; i < static_cast<int32_t>(ioData.tracks.size()); ++i) {
+			for (int32_t i = 0; i < static_cast<int32_t>(ioData.tracks.size());
+			     ++i) {
 				const SequenceTrackAssetData& track = ioData.tracks[i];
 				if (
 					track.trackType == SEQUENCE_TRACK_TYPE::TRANSFORM &&
@@ -549,11 +583,11 @@ namespace Unnamed {
 
 		if (outTrackIndex < 0) {
 			SequenceTrackAssetData track = {};
-			track.trackId   = AllocateStableId();
-			track.name      = "Transform_" + std::to_string(entityGuid);
+			track.trackId = AllocateStableId();
+			track.name = "Transform_" + std::to_string(entityGuid);
 			track.trackType = SEQUENCE_TRACK_TYPE::TRANSFORM;
 			track.blendMode = SEQUENCE_BLEND_MODE::MODE_ABSOLUTE;
-			track.priority  = 0;
+			track.priority = 0;
 			track.bindingId = bindingId;
 			SequenceSectionAssetData section = {};
 			section.sectionId = AllocateStableId();
@@ -587,10 +621,11 @@ namespace Unnamed {
 			track.sections.emplace_back(std::move(section));
 		}
 
-		outSectionIndex = (
-			mSelection.sectionIndex >= 0 &&
-			mSelection.sectionIndex < static_cast<int32_t>(track.sections.size())
-		) ? mSelection.sectionIndex : 0;
+		outSectionIndex = mSelection.sectionIndex >= 0 &&
+		                  mSelection.sectionIndex < static_cast<int32_t>(track.
+			                  sections.size()) ?
+			                  mSelection.sectionIndex :
+			                  0;
 		if (track.sections[outSectionIndex].sectionId == 0) {
 			track.sections[outSectionIndex].sectionId = AllocateStableId();
 		}
